fixes and refactors - fixed recursive reference - moved drum interfaces to types.ts - exposed DrumParameterUtils
