+++ src/midi/exports.ts
@@ -7,3 +7,4 @@ export type { ModifyMIDIOptions } from "./midi_tools/modify_midi";
 export { MIDIBuilder } from "./midi_tools/midi_builder";
 export { MIDITrack } from "./midi_track";
 export { MIDIUtils } from "./midi_tools/midi_utils";
+export { DrumParameterUtils } from "./drum_parameters";
